backend/requirements.txt
pydantic-ai
pydantic-ai-slim[duckduckgo]
fastapi
uvicorn
pydantic
httpx
beautifulsoup4
python-multipart
python-dotenv
PyPDF2
pymupdf
google-genai==1.28.0
ollama
tqdm
pymilvus
pymilvus[milvus_lite]